# GYMaster-API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GYMaster",
    "version": "v1"
  },
  "paths": {
...
    "/api/Productos/Mostrar": {
      "get": {
        "tags": [
          "Productos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioGimnasios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductosDTO"
              }
            }
          }
        }
      }
    },
    "/api/Productos/Catalogo": {
      "get": {
        "tags": [
          "Productos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioGimnasios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductosDTO"
              }
            }
          }
        }
      }
    },
    "/api/Productos/Registrar": {
      "post": {
        "tags": [
          "Productos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductosDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Productos/Modificar": {
      "put": {
        "tags": [
          "Productos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "NombreProductos",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductosDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Suscripciones/Buscar": {
      "get": {
        "tags": [
          "Suscripciones"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioUsuarios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SuscripcionesDTO"
            }
          }
        }
      }
    },
    "/api/Suscripciones/Registrar": {
      "post": {
        "tags": [
          "Suscripciones"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SuscripcionesDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Suscripciones/Modificar": {
      "put": {
        "tags": [
          "Suscripciones"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SuscripcionesDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProductosDTO": {
      "type": "object",
      "properties": {
        "usuarioGimnasios": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "stock": {
          "format": "int32",
          "type": "integer"
        },
        "idEstatus": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "SuscripcionesDTO": {
      "type": "object",
      "properties": {
        "usuarioUsuarios": {
          "type": "string"
        },
        "usuarioGimnasios": {
          "type": "string"
        },
        "tipoMembresia": {
          "type": "string"
        },
        "fechaRenovacion": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
  }
}